Capabilities
A layer can specify capabilities to determine if it should be included.
A watch must match all specified capabilities to include the layer.

See pebble/clay for explanation of capabilities and a list of available
capabilities. Note that PLATFORM_EMERY and screen size capabilities are
not included since those will never exist for real hardware.